[irep2] splitted templates specialization
