+++ src/irep2/templates/irep2_templates_expr_extra.cpp
@@ -1,0 +1,8 @@
+#include <irep2/irep2_templates_expr.h>
+
+
+expr_typedefs_empty(unknown, expr2t);
+expr_typedefs_empty(invalid, expr2t);
+expr_typedefs_empty(null_object, expr2t);
+
+
